# cmd/app/swagger-gen/swagger.yaml
openapi: 3.0.1
info:
  contact: {}
  description: "REST API for e-commerce (products, users, orders)"
  title: R2 Ecoomerce API
  version: "1.0"
servers:
- url: /v1
paths:
...
  /orders/{id}:
    get:
      description: Get order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get order
      tags:
      - Orders
  /orders/{id}/status:
    put:
      description: Update status for an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_order_adapters_http.updateStatusRequest"
        description: Status input
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Update order status
      tags:
      - Orders
      x-codegen-request-body-name: body
  /products:
    get:
      description: List products with optional filters
      parameters:
      - description: Category
        in: query
        name: category
        schema:
          type: string
      - description: Name contains
        in: query
        name: name
        schema:
          type: string
      - description: "Sort by (name, price_cents, etc)"
        in: query
        name: sort
        schema:
          type: string
      - description: asc|desc
        in: query
        name: order
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: List products
      tags:
      - Products
    post:
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_product_adapters_http.createProductRequest"
        description: Product input
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/r2-challenge_internal_product_domain.Product"
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Create product
      tags:
      - Products
      x-codegen-request-body-name: product
  /products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Delete product
      tags:
      - Products
    get:
      description: Get a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get product
      tags:
      - Products
    put:
      description: Update a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_product_adapters_http.updateProductRequest"
        description: Product input
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/r2-challenge_internal_product_domain.Product"
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Update product
      tags:
      - Products
      x-codegen-request-body-name: product
...
  /users/{id}/orders:
    get:
      description: List orders for the authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: List my orders
      tags:
      - Orders
...
components:
  schemas:
...
    internal_order_adapters_http.updateStatusRequest:
      properties:
        status:
          type: string
      required:
      - status
      type: object
    internal_product_adapters_http.createProductRequest:
      properties:
        category:
          type: string
        description:
          type: string
        inventory:
          minimum: 0
          type: integer
        name:
          minLength: 3
          type: string
        price_cents:
          minimum: 0
          type: integer
      required:
      - category
      - name
      - price_cents
      type: object
    internal_product_adapters_http.updateProductRequest:
      properties:
        category:
          type: string
        description:
          type: string
        inventory:
          minimum: 0
          type: integer
        name:
          minLength: 3
          type: string
        price_cents:
          minimum: 0
          type: integer
      required:
      - category
      - name
      - price_cents
      type: object
...
    r2-challenge_internal_product_domain.Product:
      example:
        updated_at: updated_at
        name: name
        created_at: created_at
        description: description
        price_cents: 0
        id: id
        category: category
        inventory: 0
      properties:
        category:
          type: string
        created_at:
          type: string
        description:
          type: string
        id:
          type: string
        inventory:
          minimum: 0
          type: integer
        name:
          minLength: 3
          type: string
        price_cents:
          minimum: 0
          type: integer
        updated_at:
          type: string
      required:
      - category
      - name
      - price_cents
      type: object
...